+++ tests/infer_missing_types_test/src/001_vaguer_comment_return.php
@@ -1,4 +1,11 @@
 <?php
+
+namespace NS1;
+
+use ArrayAccess;
+use ArrayObject;
+use stdClass;
+
 /** @return mixed */
 function test1() { return new ArrayObject(); }
 
@@ -9,4 +16,44 @@ function test2() {
     $instance = new ArrayObject();
   }
   return $instance;
+}
+function test3(): ArrayAccess
+{
+    return new ArrayObject();
+}
+
+/**
+ * @param object $o
+ * @return mixed
+ */
+function test4($o)
+{
+    return $o;
+}
+
+class Methods {
+
+    /**
+     * @param stdClass $o
+     * @return array TODO: Phan fails to emit PhanPluginMoreSpecificActualReturnTypeContainsFQSEN for (at)return object[]
+     */
+    function test5($o)
+    {
+        return [$o];
+    }
+    /**
+     * @param stdClass $o
+     * @return object[]
+     */
+    function test6($o)
+    {
+        return [$o];
+    }
+    /**
+     * @return iterable<object>
+     */
+    function test7()
+    {
+        return [$this];
+    }
 }